add UserInfo component
