+++ src/components/Header/Header.js
@@ -1,8 +1,7 @@
 import { useState, useEffect } from 'react';
 import { useMedia } from 'react-use';
-
 import { useLocation } from 'react-router-dom';
-
+import { UserInfo } from 'components/UserInfo/UserInfo';
 import {
   Container,
   Image,
@@ -11,17 +10,12 @@ import {
   Motivation,
   Accent,
   Svg,
-  Name,
-  Avatar,
-  AvatarLetter,
   Input,
   Button,
   Burger,
 } from './Header.styled';
-
-import goose from './goose.png';
-
-import icon from './icons.svg';
+import goose from 'images/icons/goose.png';
+import icon from 'images/sprite.svg';
 
 export default function Header() {
   const [darkTheme, setDarkTheme] = useState(
@@ -31,14 +25,10 @@ export default function Header() {
   useEffect(() => {
     localStorage.setItem('darkTheme', darkTheme);
   }, [darkTheme]);
+
   const location = useLocation();
   const isActivePageCalendar = location.pathname.includes('calendar');
-
   const isTabletOrMobile = useMedia('(max-width: 1439px)');
-
-  const name = 'Name';
-
-  const firstLetter = name.trim().slice(0, 1).toUpperCase();
 
   return (
     <Container>
@@ -74,10 +64,7 @@ export default function Header() {
             <use href={icon + (darkTheme ? '#icon-sun' : '#icon-moon')}></use>
           </Svg>
         </label>
-        <Name>{name}</Name>
-        <Avatar>
-          <AvatarLetter>{firstLetter}</AvatarLetter>
-        </Avatar>
+        <UserInfo />
       </Wrapper>
     </Container>
   );
