+++ src/components/Header/Header.styled.js
@@ -5,7 +5,7 @@ export const Container = styled.div`
   justify-content: space-between;
   align-items: center;
   margin-bottom: 64px;
-  @media screen and (min-width: 1440px) {
+  @media screen and (min-width: 1439px) {
     margin-bottom: 32px;
   }
 `;
@@ -26,7 +26,7 @@ export const Burger = styled.svg`
   stroke: #343434;
   fill: none;
   cursor: pointer;
-  @media screen and (min-width: 768px) {
+  @media screen and (min-width: 767px) {
     width: 34px;
     height: 34px;
   }
@@ -59,58 +59,7 @@ export const Accent = styled.span`
   color: #3e85f3;
 `;
 
-//UserInfo
-
-export const Name = styled.span`
-  margin-right: 14px;
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 18px;
-  color: #343434;
-  @media screen and (min-width: 768px) {
-    font-size: 18px;
-  }
-  @media screen and (min-width: 1440px) {
-  }
-`;
-
-export const Avatar = styled.div`
-  width: 32px;
-  height: 32px;
-  display: flex;
-  align-items: center;
-  border-radius: 50%;
-  border: 1.8px solid #3e85f3;
-  background-color: #f7f6f9;
-  justify-content: center;
-  @media screen and (min-width: 768px) {
-    width: 44px;
-    height: 44px;
-  }
-  @media screen and (min-width: 1440px) {
-  }
-`;
-
-export const AvatarLetter = styled.span`
-  display: flex;
-  align-items: center;
-  font-style: normal;
-  font-weight: 700;
-  font-size: 26px;
-  color: #3e85f3;
-  text-shadow: 0px 47px 355px rgba(0, 0, 0, 0.07),
-    0px 9.4px 57.6875px rgba(0, 0, 0, 0.035);
-  @media screen and (min-width: 768px) {
-    font-weight: 700;
-    font-size: 32px;
-  }
-  @media screen and (min-width: 1440px) {
-  }
-`;
-
-//Toggle
+//Theme Toggle
 
 export const Input = styled.input`
   position: absolute;
@@ -132,7 +81,7 @@ export const Svg = styled.svg`
   margin-right: 14px;
   fill: #3e85f3;
   cursor: pointer;
-  @media screen and (min-width: 768px) {
+  @media screen and (min-width: 767px) {
     width: 26px;
     height: 26px;
   }
@@ -142,6 +91,11 @@ export const Svg = styled.svg`
 export const Button = styled.button`
   width: 130px;
   height: 42px;
+  font-family: 'Inter';
+  font-style: normal;
+  font-weight: 600;
+  font-size: 14px;
+  line-height: 18px;
   color: #fff;
 
   background: #3e85f3;
